+++ satella/time.py
@@ -1,13 +1,13 @@
 import copy
 import inspect
-import math
 import time
 import typing as tp
 import warnings
 from concurrent.futures import Future
 from functools import wraps  # import from functools to prevent circular import exception
 
-__all__ = ['measure', 'time_as_int', 'time_ms', 'sleep', 'time_us', 'ExponentialBackoff']
+__all__ = ['measure', 'time_as_int', 'time_ms', 'sleep', 'time_us', 'ExponentialBackoff',
+           'parse_time_string']
 
 from satella.exceptions import WouldWaitMore
 
@@ -364,3 +364,39 @@ class ExponentialBackoff:
         Called when something successes.
         """
         self.counter = self.start
+
+
+TIME_MODIFIERS = [
+    ('s', 1),
+    ('m', 60),
+    ('h', 60*60),
+    ('d', 24*60*60),
+    ('w', 7*24*60*60)
+]
+
+
+def parse_time_string(s: tp.Union[int, float, str]) -> float:
+    """
+    Parse a time string into seconds, so eg. '30m' will be equal to 1800, and so will
+    be '30 min'.
+
+    This will correctly parse:
+    - seconds
+    - minutes
+    - hours
+    - days
+    - weeks
+
+    .. warning:: This does not handle fractions of a second!
+
+    :param s: time string or time value in seconds
+    :return: value in seconds
+    """
+    if isinstance(s, (int, float)):
+        return s
+
+    for modifier, multiple in TIME_MODIFIERS:
+        if modifier in s:
+            return float(s[:s.index(modifier)]) * multiple
+
+    return float(s)