+++ satella/instrumentation/metrics/structures/threadpool.py
@@ -100,8 +100,8 @@ class MetrifiedThreadPoolExecutor(ThreadPoolExecutor):
         self._idle_semaphore = threading.Semaphore(0)
         self._broken = False
         if not hasattr(self, '_thread_name_prefix'):
-            self._thread_name_prefix = (thread_name_prefix or
-                                        ("ThreadPoolExecutor-%d" % self._counter()))
+            self._thread_name_prefix = thread_name_prefix or ("ThreadPoolExecutor-%d" %
+                                                              self._counter())
         self.waiting_time_metric = time_spent_waiting or EmptyMetric('')
         self.executing_time_metric = time_spent_executing or EmptyMetric('')
         self.metric_level = metric_level