+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.46
 
 * added kwargs to ThreadCollection.from_class
+* fixed a bug in CustomException