+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added kwargs to ThreadCollection.from_class
 * fixed a bug in CustomException
+* added `parse_time_string`