+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.14.47
 
+* more time-related calls will accept time strings