+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.47
 
 * more time-related calls will accept time strings
+* added optional `delay` argument to `call_in_separate_thread`