+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.48
+
+* **bugfix** fixed `CPUProfileBuilderThread`