+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# v2.14.48
+# v2.14.49
 
 * **bugfix** fixed `CPUProfileBuilderThread`